requirements.txt
click==7.1.2
cycler==0.10.0
et-xmlfile==1.0.1
Flask==1.1.2
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
joblib==0.17.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.3
openpyxl==3.0.5
pandas==1.1.4
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.4
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
text-unidecode==1.3
threadpoolctl==2.1.0
Werkzeug==1.0.1
xgboost==0.90
xlrd==1.2.0
